+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/txClient/TerminologyClientR3.java
@@ -223,4 +223,10 @@ public class TerminologyClientR3 implements ITerminologyClient {
     client.setContentLanguage(lang);
     return this;
   }
+  
+  @Override
+  public int getUseCount() {
+    return client.getUseCount();
+  }
+
 }