+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/txClient/TerminologyClientR4.java
@@ -235,4 +235,11 @@ public class TerminologyClientR4 implements ITerminologyClient {
     client.setContentLanguage(lang);
     return this;
   }
+  
+  @Override
+  public int getUseCount() {
+    return client.getUseCount();
+  }
+
+  
 }