+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/context/SimpleWorkerContext.java
@@ -68,6 +68,9 @@ import org.hl7.fhir.r5.profilemodel.PEBuilder;
 import org.hl7.fhir.r5.terminologies.CodeSystemUtilities;
 import org.hl7.fhir.r5.terminologies.JurisdictionUtilities;
 import org.hl7.fhir.r5.terminologies.client.ITerminologyClient;
+import org.hl7.fhir.r5.terminologies.client.TerminologyClientManager;
+import org.hl7.fhir.r5.terminologies.client.TerminologyClientManager.ITerminologyClientFactory;
+import org.hl7.fhir.r5.terminologies.client.TerminologyClientR5;
 import org.hl7.fhir.r5.utils.validation.IResourceValidator;
 import org.hl7.fhir.r5.utils.R5Hacker;
 import org.hl7.fhir.r5.utils.XVerExtensionManager;
@@ -248,6 +251,7 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
     public SimpleWorkerContext fromPackage(NpmPackage pi) throws IOException, FHIRException {
       SimpleWorkerContext context = getSimpleWorkerContextInstance();
       context.setAllowLoadingDuplicates(allowLoadingDuplicates);
+      context.tcc = new TerminologyClientManager(TerminologyClientR5.factory());
       context.loadFromPackage(pi, null);
       return build(context);
     }
@@ -256,6 +260,7 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
       SimpleWorkerContext context = getSimpleWorkerContextInstance();
       context.setAllowLoadingDuplicates(allowLoadingDuplicates);      
       context.version = pi.getNpm().asString("version");
+      context.tcc.setFactory(loader.txFactory());
       context.loadFromPackage(pi, loader);
       context.finishLoading(genSnapshots);
       return build(context);
@@ -327,22 +332,23 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
       loadBytes(name, stream);
   }
 
-  public String connectToTSServer(ITerminologyClient client, String log) {
+  public String connectToTSServer(ITerminologyClientFactory factory, ITerminologyClient client, String log) {
     try {
       txLog("Connect to "+client.getAddress());
-      tcc.setClient(client);
+      tcc.setFactory(factory);
+      tcc.setMasterClient(client);
       if (log != null && (log.endsWith(".htm") || log.endsWith(".html"))) {
         txLog = new HTMLClientLogger(log);
       } else {
         txLog = new TextClientLogger(log);
       }
-      tcc.getClient().setLogger(txLog);
-      tcc.getClient().setUserAgent(userAgent);
+      tcc.setLogger(txLog);
+      tcc.setUserAgent(userAgent);
 
-      final CapabilityStatement capabilitiesStatementQuick = txCache.hasCapabilityStatement() ? txCache.getCapabilityStatement() : tcc.getClient().getCapabilitiesStatementQuick();
+      final CapabilityStatement capabilitiesStatementQuick = txCache.hasCapabilityStatement() ? txCache.getCapabilityStatement() : tcc.getMasterClient().getCapabilitiesStatementQuick();
       txCache.cacheCapabilityStatement(capabilitiesStatementQuick);
 
-      final TerminologyCapabilities capabilityStatement = txCache.hasTerminologyCapabilities() ? txCache.getTerminologyCapabilities() : tcc.getClient().getTerminologyCapabilities();
+      final TerminologyCapabilities capabilityStatement = txCache.hasTerminologyCapabilities() ? txCache.getTerminologyCapabilities() : tcc.getMasterClient().getTerminologyCapabilities();
       txCache.cacheTerminologyCapabilities(capabilityStatement);
 
       setTxCaps(capabilityStatement);
@@ -531,6 +537,9 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
 	    if (version.equals("current")) {
 	      version = "5.0.0";
 	    }
+	  }
+	  if (loader != null) {
+	    tcc.setFactory(loader.txFactory());
 	  }
 	  return t;
 	}