+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/terminologies/client/TerminologyClientContext.java
@@ -1,65 +1,70 @@
 package org.hl7.fhir.r5.terminologies.client;
 
+import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
 import org.hl7.fhir.r5.model.CanonicalResource;
 import org.hl7.fhir.r5.model.TerminologyCapabilities;
-import org.hl7.fhir.r5.model.ValueSet;
+import org.hl7.fhir.r5.terminologies.client.TerminologyClientContext.TerminologyClientContextUseCount;
 
 public class TerminologyClientContext {
-  private String cacheId;
-  private boolean isTxCaching;
-  private int serverQueryCount = 0;
-  private final Set<String> cached = new HashSet<>();
-  protected String server;
+  public class TerminologyClientContextUseCount {
+    private int expands;
+    private int validates;
+    public int getExpands() {
+      return expands;
+    }
+    public void setExpands(int expands) {
+      this.expands = expands;
+    }
+    public int getValidates() {
+      return validates;
+    }
+    public void setValidates(int validates) {
+      this.validates = validates;
+    }
+    
+  }
+
   private ITerminologyClient client;
+  Map<String, TerminologyClientContextUseCount> useCounts = new HashMap<>();
   private TerminologyCapabilities txcaps;
+  private final Set<String> cached = new HashSet<>();
+  private boolean master;
 
-  public String getCacheId() {
-    return cacheId; 
-  }
-  
-  public void setCacheId(String cacheId) {
-    this.cacheId = cacheId;
-  }
-  
-  public boolean isTxCaching() {
-    return isTxCaching;
-  }
-  
-  public void setTxCaching(boolean isTxCaching) {
-    this.isTxCaching = isTxCaching;
-  }
-  
-  public int getServerQueryCount() {
-    return serverQueryCount;
-  }
-  
-  public void setServerQueryCount(int serverQueryCount) {
-    this.serverQueryCount = serverQueryCount;
-  }
-  
-  public Set<String> getCached() {
-    return cached;
+  protected TerminologyClientContext(ITerminologyClient client, boolean master) {
+    super();
+    this.client = client;
+    this.master = master;
   }
 
-  public String getServer() {
-    return server;
+  public Map<String, TerminologyClientContextUseCount> getUseCounts() {
+    return useCounts;
   }
 
-  public void setServer(String server) {
-    this.server = server;
+  public boolean isMaster() {
+    return master;
   }
 
   public ITerminologyClient getClient() {
     return client;
   }
 
-  public void setClient(ITerminologyClient client) {
-    this.client = client;
+  public void seeUse(String s, boolean expand) {
+    TerminologyClientContextUseCount uc = useCounts.get(s);
+    if (uc == null) {
+      uc = new TerminologyClientContextUseCount();
+      useCounts.put(s,uc);
+    }
+    if (expand) {
+      uc.expands++;
+    } else {
+      uc.validates++;
+    }
   }
-    
+
   public TerminologyCapabilities getTxcaps() {
     return txcaps;
   }
@@ -68,18 +73,8 @@ public class TerminologyClientContext {
     this.txcaps = txcaps;
   }
 
-  public void copy(TerminologyClientContext other) {
-    cacheId = other.cacheId;  
-    isTxCaching = other.isTxCaching;
-    cached.addAll(other.cached);
-//    tsServer = other.tsServer;
-    client = other.client;
-    txcaps = other.txcaps;
-
-  }
-
-  public boolean usingCache() {
-    return isTxCaching && cacheId != null;
+  public Set<String> getCached() {
+    return cached;
   }
 
   public boolean alreadyCached(CanonicalResource cr) {
@@ -90,6 +85,12 @@ public class TerminologyClientContext {
     cached.add(cr.getVUrl());
   }
 
-  
-  
+  public String getAddress() {
+    return client.getAddress();
+  }
+
+  public int getUseCount() {
+    return getClient().getUseCount();
+  }
+
 }